VideoPlayer: extract the C++ extension
Now, VideoPlayer is a dependency to make release easy. The layer
remains. Cython is not necessary anymore.
