+++ fofix/core/videolayer.py
@@ -1,9 +1,9 @@
 #####################################################################
-# -*- coding: iso-8859-1 -*-                                        #
+# -*- coding: utf-8 -*-                                             #
 #                                                                   #
 # Frets on Fire X (FoFiX)                                           #
-# Copyright (C) 2010 FoFiX Team                                     #
-#               2010 John Stumpo                                    #
+# Copyright (C) 2010 John Stumpo                                    #
+#               2010-2020 FoFiX Team                                #
 #                                                                   #
 # This program is free software; you can redistribute it and/or     #
 # modify it under the terms of the GNU General Public License       #
@@ -21,91 +21,22 @@
 # MA  02110-1301, USA.                                              #
 #####################################################################
 
-# First a thin wrapper around VideoPlayer from graphics.VideoPlayer.c...
-
-cdef extern from "VideoPlayer.h":
-    ctypedef struct CVideoPlayer "VideoPlayer":
-        pass
-
-    ctypedef int GQuark
-    ctypedef struct GError:
-        GQuark domain
-        char* message
-    GQuark G_FILE_ERROR
-    GQuark VIDEO_PLAYER_ERROR
-    void g_error_free(GError*)
-
-    CVideoPlayer* video_player_new(char*, GError**)
-    void video_player_destroy(CVideoPlayer*)
-    void video_player_play(CVideoPlayer*)
-    void video_player_pause(CVideoPlayer*)
-    bint video_player_advance(CVideoPlayer*, double, GError**)
-    bint video_player_bind_frame(CVideoPlayer*, GError**)
-    bint video_player_eof(CVideoPlayer*)
-    double video_player_aspect_ratio(CVideoPlayer*)
-
-class VideoPlayerError(Exception):
-    pass
-
-cdef object raise_from_gerror(GError* err):
-    assert err is not NULL
-    if err.domain == VIDEO_PLAYER_ERROR:
-        exc = VideoPlayerError(err.message)
-    elif err.domain == G_FILE_ERROR:
-        exc = IOError(err.message)
-    else:
-        exc = Exception(err.message)
-    g_error_free(err)
-    raise exc
-
-cdef class VideoPlayer(object):
-    cdef CVideoPlayer* player
-
-    def __cinit__(self, char* filename):
-        cdef GError* err = NULL
-        self.player = video_player_new(filename, &err)
-        if self.player is NULL:
-            raise_from_gerror(err)
-
-    def __dealloc__(self):
-        if self.player is not NULL:
-            video_player_destroy(self.player)
-
-    def play(self):
-        video_player_play(self.player)
-
-    def pause(self):
-        video_player_pause(self.player)
-
-    def advance(self, double newpos):
-        cdef GError* err = NULL
-        if not video_player_advance(self.player, newpos, &err):
-            raise_from_gerror(err)
-
-    def bind_frame(self):
-        cdef GError* err = NULL
-        if not video_player_bind_frame(self.player, &err):
-            raise_from_gerror(err)
-
-    def eof(self):
-        return video_player_eof(self.player)
-
-    def aspect_ratio(self):
-        return video_player_aspect_ratio(self.player)
+""" A layer for playing a video """
 
 
-# And now, a layer for playing a video.
+import numpy as np
+from OpenGL import GL as gl
+from videoplayer._VideoPlayer import VideoPlayer
+from videoplayer._VideoPlayer import VideoPlayerError
 
 from fofix.core.View import BackgroundLayer
 from fofix.core.Input import KeyListener
-import numpy as np
 from fofix.core import cmgl
 
-# This is Cython after all, so we may as well directly bind to OpenGL for the video layer implementation.
-include "gl.pxi"
 
 class VideoLayer(BackgroundLayer, KeyListener):
-    def __init__(self, engine, filename, mute = False, loop = False, cancellable = False):
+
+    def __init__(self, engine, filename, mute=False, loop=False, cancellable=False):
         self.engine = engine
         self.filename = filename
         self.mute = mute  # TODO: audio
@@ -156,12 +87,12 @@ class VideoLayer(BackgroundLayer, KeyListener):
 
         self.player.bind_frame()
 
-        glMatrixMode(GL_PROJECTION)
-        glPushMatrix()
-        glLoadIdentity()
-        glMatrixMode(GL_MODELVIEW)
-        glPushMatrix()
-        glLoadIdentity()
+        gl.glMatrixMode(gl.GL_PROJECTION)
+        gl.glPushMatrix()
+        gl.glLoadIdentity()
+        gl.glMatrixMode(gl.GL_MODELVIEW)
+        gl.glPushMatrix()
+        gl.glLoadIdentity()
 
         # Prepare to draw the video over a black rectangle that will fill the rest of the space.
         background_vertices = np.array([[ 1.0,  1.0],
@@ -178,19 +109,19 @@ class VideoLayer(BackgroundLayer, KeyListener):
                               [1.0, 1.0]], dtype=np.float32)
 
         # The black rectangle.
-        glColor4f(0.0, 0.0, 0.0, 1.0)
-        cmgl.draw_arrays(GL_QUADS, vertices=background_vertices)
+        gl.glColor4f(0.0, 0.0, 0.0, 1.0)
+        cmgl.draw_arrays(gl.GL_QUADS, vertices=background_vertices)
 
         # The actual video.
-        glEnable(GL_TEXTURE_2D)
-        glColor4f(1.0, 1.0, 1.0, 1.0)
-        cmgl.draw_arrays(GL_QUADS, vertices=video_vertices, texcoords=texcoords)
-        glDisable(GL_TEXTURE_2D)
+        gl.glEnable(gl.GL_TEXTURE_2D)
+        gl.glColor4f(1.0, 1.0, 1.0, 1.0)
+        cmgl.draw_arrays(gl.GL_QUADS, vertices=video_vertices, texcoords=texcoords)
+        gl.glDisable(gl.GL_TEXTURE_2D)
 
-        glPopMatrix()
-        glMatrixMode(GL_PROJECTION)
-        glPopMatrix()
-        glMatrixMode(GL_MODELVIEW)
+        gl.glPopMatrix()
+        gl.glMatrixMode(gl.GL_PROJECTION)
+        gl.glPopMatrix()
+        gl.glMatrixMode(gl.GL_MODELVIEW)
 
         if self.player.eof():
             if self.loop:
