+++ requirements.txt
@@ -1,5 +1,4 @@
 Cerealizer==0.8.2
-Cython==0.27.3
 numpy==1.13.3
 olefile==0.44
 Pillow==4.3.0
@@ -9,3 +8,4 @@ PyOpenGL==3.1.0
 PyOpenGL-accelerate==3.1.0
 fretwork==0.4.0
 pypitch==1.2
+videoplayer==1.0
